Atualização

+++ Batch_senac/src/main/java/br/com/batch/senac/dto/DependentesDto.java
@@ -2,13 +2,13 @@ package br.com.batch.senac.dto;
 
 public class DependentesDto {
 	
-	private Long nro_dependentes_ir;
+	private Integer nro_dependentes_ir;
 
-	public Long getNro_dependentes_ir() {
+	public Integer getNro_dependentes_ir() {
 		return nro_dependentes_ir;
 	}
 
-	public void setNro_dependentes_ir(Long nro_dependentes_ir) {
+	public void setNro_dependentes_ir(Integer nro_dependentes_ir) {
 		this.nro_dependentes_ir = nro_dependentes_ir;
 	}
